requirements.txt
fal-client>=0.5.0
requests>=2.28.0
numpy<2.0.0
Pillow>=9.0.0
torch>=1.9.0
